support options for int:i/2, int:ni/2 and int:interpretable/2
Searching for source code for a beam file, or the other way around,
requires knowledge of the structure of the project. Before, it was
assumed that the structure was the standard one, but even that would
fail for the "test" directory.

Supported options are {ebin_dir, Dir} and {src_dirs, [Dir]}, other
values are ignored.